refactor: rename application api view for easier editor search
